+++ test/test_helper.rb
@@ -17,11 +17,10 @@ require "mocha/minitest"
 
 require "shrine"
 
-class Shrine
-  def warn(*); end # disable mime_type warnings
-end
+Shrine.logger = Logger.new(nil) # disable mime_type warnings
 
 Mocha::Configuration.prevent(:stubbing_non_existent_method)
 
 require "./test/support/generic_helper"
 require "./test/support/deprecated_helper"
+require "./test/support/logging_helper"